Broke out DimView tests. Added Equals test. Found bug. Fixed.
